Fix wrong fields, wrong callees and wrong math in the AmtBar and GreatPower families
Nine confirmed semantic defects found by reviewing TGreatPower and TAmtBar against
the disassembly. All vtables were already 100%, so none of these were visible to
the vtable or policy gates.

TAmtBar / trade-cluster family:

- UpdateBarValuesAndRefresh (0x00588630) assigned both arguments to the wrong
  fields: arg1 belongs at +0x60 and arg2 at +0x62 (0x00588640/0x00588646), and the
  body had them crossed. Every amount bar in the trade/rail/industry/shipyard
  clusters sets its range and current value through SetBarMetric -> here, so both
  were swapped at runtime.
- DoMouseCommand (0x00588950) computed point.x / (width * segments) where the
  FILD/FMULP/FDIVP chain at 0x00588975-0x00588987 computes point.x * segments /
  width; the ported form collapses to ~0 for realistic clicks. Also inlined the
  ClampAndApplyTradeMoveValue helper (the original is one 173-byte body), replaced
  the GetWindow() virtual with the ownerContext field load the original does, and
  dropped a raw *(double*)0x006631a0 read of the -1.0 constant-pool entry.
- QueryValue was a fabricated stand-in: all 14 call sites dispatch slot 0x1e8,
  which is past TAmtBar's 0x1a8 vtable end and is
  TNumberText::UpdateControlCachedIntFromWindowText (slot 0x7a). On a real
  TNumberText, QueryValue read +0x62 (inside TEditText) instead of value at +0xa0,
  so every cluster gating on a move/sell quantity compared garbage. Retyped the
  Sell/move/avai/mCap controls to TNumberText across TTradeCluster, TRailCluster,
  TIndustryCluster, TShipyardCluster, TAmtBarCluster and TAmtBar, and deleted
  TAmtBar::QueryValue. Verified per site against listings at 0x00586e70,
  0x00588950, 0x00589da0, 0x00588ff0, 0x0058a940, 0x005873e0, 0x00587900,
  0x00587950. Reopens the premature close of bd ktqe, whose fix was not on main.
- TTraderAmtBar::ApplyMoveClamp (0x0058b070) divided frameHeight38 instead of
  frameWidth34 (+0x34 at 0x0058b09f) and added a tradeCapacity != 0 guard the
  binary does not have.
- ApplyMoveClamp's second parameter is a short, not an int (MOV AX,[ESP+4] at the
  base, MOV DI,word ptr [ESP+0x14] in the override); base and override updated
  together.

TGreatPower family:

- ReleaseProposalQueueSlot7F (0x004df5a0) loads +0x848 = turnEventQueue, not
  proposalQueue (+0x84c, which is what 0x004df580 loads). The port collapsed both
  onto proposalQueue, so the turn-event queue was never cleared and the proposal
  queue was cleared twice. Also dropped a null check 0x004df580 does not have.
- ApplyRelationDeltaToCityStockAndUpdateState1f4 (0x004dcca0) used
  relationDeltaCurrent (+0x16a) where the loop base ESI-0x5c resolves to
  relationDeltaSnapshot (+0x198) - three sites in one loop over two adjacent,
  identically shaped short[0x17] arrays.
- TCountry's diplomacy deserializer read into raw offsets 0x94/0xc2/0xf0, each
  0x1e low against the current layout; 0x94 is foreignMinister, so the first
  ReadBytes overwrote all three minister pointers. Each read was also paired with
  a byte-swap of a different array. Now reads and swaps the same members as its
  write twin.

Elsewhere:

- GetTickCountDiv16 (0x00493250) called KERNEL32 GetTickCount; the original calls
  the WINMM timeGetTime import at 0x006ab5fc. winmm was already linked and
  ui_timing.cpp already used timeGetTime() >> 4, so this was the outlier.
- GetActiveNationState() was an unmarked helper bundling
  g_apNationStates[g_pSimMgr->GetActiveNationId()], which all four callers inline
  (e.g. 0x0058abf9). Inlined and deleted, along with its prelude declaration.
- TStream::IsAtEnd: sequenced GetPosition() into a local so VC5 emits slot 0x28
  before slot 0x30 as the original does, instead of leaving it to unspecified
  operand evaluation order.
- Replaced raw char*+offset pokes with the typed TGreatPower members they resolve
  to (city +0x894, trackedObjectList +0x89c, serializedStatusFlags +0x8c8) and
  dropped redundant casts in TShipyardCluster::DoEvent and TTraderAmtBar.

The one item deliberately not changed: the reinterpret_cast<int>(TZone*) in
ComputeMapActionContextCompositeScoreForNation. Reading 0x004e9060 showed the
multi-candidate path reloads the same [ESP+0x38] slot, so this is a stack-slot
liveness question, not a one-line fix; the misleading "original artifact" comment
is replaced with an UNRESOLVED_FIELD_ATTRIBUTION tag recording both readings.

Verification: just build, just gates (all pass), just test, just stats.
Exact functions 4051 -> 4059 (+8); exact/original 50.69% -> 50.79%; average
similarity 77.72% -> 77.76%. To 100%: 0x00588630, 0x00493250, 0x004df580,
0x004df5a0, 0x004dcca0, 0x0058abf0, 0x0058b070. 0x00588950 5.88% -> 77.17%,
0x0058af80 42.52% -> 55.81%, 0x00589260 34.15% -> 48.00%.
The 7 remaining sub-baseline functions are untouched code; all triage as
inconclusive or effective (TStream::IsAtEnd now proves effective at 87.50%).

bd 4km.8 4km.9 4km.10 4km.11 ktqe 1uj.28.5 1uj.28.6 1uj.28.8 ol2y v25k pnc0

+++ include/game/ui_widgets/TTraderAmtBar.h
@@ -14,7 +14,7 @@ public:
   // ~TTraderAmtBar is compiler-generated (implicit virtual dtor).
   DECLARE_DYNCREATE(TTraderAmtBar)
   void DoPostCreate(int arg) override;
-  short ApplyMoveClamp(int baseValue, int requestedValue) override;
+  short ApplyMoveClamp(int baseValue, short requestedValue) override;
   void RenderPrimarySurfaceOverlayPanelWithClipCache() override;
 
   void UpdateFromScaleOrRatio(int scaleValue, int ratioValue);
